chore: 엔티티 파일 이름 변경 및 파일 정돈
- BaseEntity와 UpdateEntity를 extend하고, 필드에서는 제거
