--- src/main/java/com/example/RealMatch/brand/domain/entity/BrandCategoryEntity.java
@@ -1,47 +1,0 @@
-package com.example.RealMatch.brand.domain.entity;
-
-import java.time.LocalDateTime;
-
-import com.example.RealMatch.global.common.BaseEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "p_brand_category")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class BrandCategoryEntity extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 100)
-    private String name;
-
-    @Column(name = "is_deleted")
-    private Boolean isDeleted;
-
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-
-    @Builder
-    public BrandCategoryEntity(String name) {
-        this.name = name;
-        this.isDeleted = false;
-    }
-
-    public void softDelete() {
-        this.isDeleted = true;
-        this.deletedAt = LocalDateTime.now();
-    }
-}
