+++ src/main/java/com/example/RealMatch/brand/domain/entity/Tag.java
@@ -1,9 +1,9 @@
-package com.example.RealMatch.campaign.domain.entity;
+package com.example.RealMatch.brand.domain.entity;
 
 import java.time.LocalDateTime;
 
 import com.example.RealMatch.global.common.BaseEntity;
-import com.example.RealMatch.user.domain.entity.UserEntity;
+import com.example.RealMatch.user.domain.entity.User;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -20,40 +20,47 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "p_campaign_apply")
+@Table(name = "p_tag")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CampaignApplyEntity extends BaseEntity {
+public class Tag extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private UserEntity user;
+    @JoinColumn(name = "tag_parent_id", nullable = false)
+    private BrandTagParent tagParent;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "campaign_id", nullable = false)
-    private CampaignEntity campaign;
+    @JoinColumn(name = "brand_id", nullable = false)
+    private Brand brand;
 
-    @Column(nullable = false, length = 1000)
-    private String reason;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
+    @Column(name = "category_name", nullable = false, length = 100)
+    private String categoryName;
+
+    @Column(nullable = false, length = 100)
+    private String name;
+
+    @Column(name = "is_deleted", nullable = false)
+    private boolean isDeleted;
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
 
-    @Column(name = "is_deleted")
-    private Boolean isDeleted;
-
     @Builder
-    public CampaignApplyEntity(UserEntity user, CampaignEntity campaign, String reason) {
+    public Tag(BrandTagParent tagParent, Brand brand, User user,
+               String categoryName, String name) {
+        this.tagParent = tagParent;
+        this.brand = brand;
         this.user = user;
-        this.campaign = campaign;
-        this.reason = reason;
+        this.categoryName = categoryName;
+        this.name = name;
         this.isDeleted = false;
     }
