+++ src/main/java/com/example/RealMatch/global/config/oauth/CustomOAuth2User.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 
-import com.example.RealMatch.match.domain.entity.User;
+import com.example.RealMatch.user.domain.entity.User;
 
 import lombok.Getter;
