+++ src/main/java/com/example/RealMatch/global/config/service/CustomOAuth2UserService.java
@@ -13,7 +13,7 @@ import com.example.RealMatch.global.config.oauth.CustomOAuth2User;
 import com.example.RealMatch.global.config.oauth.KakaoUserInfo;
 import com.example.RealMatch.global.config.oauth.NaverUserInfo;
 import com.example.RealMatch.global.config.oauth.OAuth2UserInfo;
-import com.example.RealMatch.match.domain.entity.User;
+import com.example.RealMatch.user.domain.entity.User;
 import com.example.RealMatch.match.domain.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
