+++ src/main/java/com/example/RealMatch/match/domain/entity/MatchCampaignHistory.java
@@ -1,11 +1,13 @@
 package com.example.RealMatch.match.domain.entity;
 
-import com.example.RealMatch.match.domain.entity.enums.TagType;
+import java.time.LocalDateTime;
+
+import com.example.RealMatch.campaign.domain.entity.Campaign;
+import com.example.RealMatch.global.common.UpdateBaseEntity;
+import com.example.RealMatch.user.domain.entity.User;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -19,26 +21,27 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "p_matching_tag")
+@Table(name = "p_match_campaign_history")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MatchingTagEntity {
+public class MatchCampaignHistory extends UpdateBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "test_id", nullable = false)
-    private MatchingTestEntity test;
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
-    @Enumerated(EnumType.STRING)
-    @Column(name = "tag_type", length = 30)
-    private TagType tagType;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "campaign_id", nullable = false)
+    private Campaign campaign;
+
 
     @Builder
-    public MatchingTagEntity(MatchingTestEntity test, TagType tagType) {
-        this.test = test;
-        this.tagType = tagType;
+    public MatchCampaignHistory(User user, Campaign campaign) {
+        this.user = user;
+        this.campaign = campaign;
     }
 }
