+++ src/main/java/com/example/RealMatch/user/domain/entity/AuthenticationMethod.java
@@ -2,6 +2,7 @@ package com.example.RealMatch.user.domain.entity;
 
 import java.time.LocalDateTime;
 
+import com.example.RealMatch.global.common.UpdateBaseEntity;
 import com.example.RealMatch.user.domain.entity.enums.AuthProvider;
 
 import jakarta.persistence.Column;
@@ -24,7 +25,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "p_authentication_methods")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AuthenticationMethodEntity {
+public class AuthenticationMethod extends UpdateBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -32,7 +33,7 @@ public class AuthenticationMethodEntity {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    private UserEntity user;
+    private User user;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
@@ -44,23 +45,15 @@ public class AuthenticationMethodEntity {
     @Column(length = 255)
     private String email;
 
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
     @Builder
-    public AuthenticationMethodEntity(UserEntity user, AuthProvider provider, String providerId, String email) {
+    public AuthenticationMethod(User user, AuthProvider provider, String providerId, String email) {
         this.user = user;
         this.provider = provider;
         this.providerId = providerId;
         this.email = email;
-        this.createdAt = LocalDateTime.now();
     }
 
     public void updateEmail(String email) {
         this.email = email;
-        this.updatedAt = LocalDateTime.now();
     }
 }
