+++ src/main/java/com/example/RealMatch/user/domain/entity/Term.java
@@ -2,6 +2,7 @@ package com.example.RealMatch.user.domain.entity;
 
 import java.time.LocalDateTime;
 
+import com.example.RealMatch.global.common.BaseEntity;
 import com.example.RealMatch.user.domain.entity.enums.TermName;
 
 import jakarta.persistence.Column;
@@ -21,7 +22,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "p_term")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TermEntity {
+public class Term extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -37,22 +38,14 @@ public class TermEntity {
     @Column(name = "is_required", nullable = false)
     private boolean isRequired;
 
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
     @Builder
-    public TermEntity(TermName name, String version, boolean isRequired) {
+    public Term(TermName name, String version, boolean isRequired) {
         this.name = name;
         this.version = version;
         this.isRequired = isRequired;
-        this.createdAt = LocalDateTime.now();
     }
 
     public void updateVersion(String version) {
         this.version = version;
-        this.updatedAt = LocalDateTime.now();
     }
 }
