--- src/main/java/com/example/RealMatch/user/domain/entity/UserEntity.java
@@ -1,105 +1,0 @@
-package com.example.RealMatch.user.domain.entity;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-import com.example.RealMatch.global.common.BaseEntity;
-import com.example.RealMatch.user.domain.entity.enums.Gender;
-import com.example.RealMatch.user.domain.entity.enums.Role;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "p_user")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserEntity extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 100)
-    private String name;
-
-    @Enumerated(EnumType.STRING)
-    @Column(length = 20)
-    private Gender gender;
-
-    private LocalDate birth;
-
-    @Column(nullable = false, length = 100)
-    private String nickname;
-
-    @Column(nullable = false, length = 255)
-    private String email;
-
-    @Column(length = 500)
-    private String address;
-
-    @Column(name = "detail_address", length = 500)
-    private String detailAddress;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 20)
-    private Role role;
-
-    @Column(name = "updated_at")
-    private LocalDate updatedAt;
-
-    @Column(name = "deleted_at")
-    private LocalDate deletedAt;
-
-    @Column(name = "deleted_by")
-    private Long deletedBy;
-
-    @Column(name = "last_login")
-    private LocalDateTime lastLogin;
-
-    @Column(name = "profile_image_url", length = 500)
-    private String profileImageUrl;
-
-    @Builder
-    public UserEntity(String name, Gender gender, LocalDate birth, String nickname,
-                      String email, String address, String detailAddress, Role role,
-                      String profileImageUrl) {
-        this.name = name;
-        this.gender = gender;
-        this.birth = birth;
-        this.nickname = nickname;
-        this.email = email;
-        this.address = address;
-        this.detailAddress = detailAddress;
-        this.role = role;
-        this.profileImageUrl = profileImageUrl;
-    }
-
-    public void updateLastLogin() {
-        this.lastLogin = LocalDateTime.now();
-    }
-
-    public void updateProfile(String name, String nickname, String address, String detailAddress, String profileImageUrl) {
-        this.name = name;
-        this.nickname = nickname;
-        this.address = address;
-        this.detailAddress = detailAddress;
-        this.profileImageUrl = profileImageUrl;
-        this.updatedAt = LocalDate.now();
-    }
-
-    public void softDelete(Long deletedBy) {
-        this.deletedAt = LocalDate.now();
-        this.deletedBy = deletedBy;
-    }
-}
