+++ src/main/java/com/example/RealMatch/user/domain/entity/UserTerm.java
@@ -2,6 +2,8 @@ package com.example.RealMatch.user.domain.entity;
 
 import java.time.LocalDateTime;
 
+import com.example.RealMatch.global.common.BaseEntity;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -20,7 +22,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "p_user_term")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserTermEntity {
+public class UserTerm extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -28,11 +30,11 @@ public class UserTermEntity {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    private UserEntity user;
+    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "term_id", nullable = false)
-    private TermEntity term;
+    private Term term;
 
     @Column(name = "is_agreed", nullable = false)
     private boolean isAgreed;
@@ -40,18 +42,11 @@ public class UserTermEntity {
     @Column(name = "agreed_at")
     private LocalDateTime agreedAt;
 
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
     @Builder
-    public UserTermEntity(UserEntity user, TermEntity term, boolean isAgreed) {
+    public UserTerm(User user, Term term, boolean isAgreed) {
         this.user = user;
         this.term = term;
         this.isAgreed = isAgreed;
-        this.createdAt = LocalDateTime.now();
         if (isAgreed) {
             this.agreedAt = LocalDateTime.now();
         }
@@ -60,12 +55,10 @@ public class UserTermEntity {
     public void agree() {
         this.isAgreed = true;
         this.agreedAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
     }
 
     public void disagree() {
         this.isAgreed = false;
         this.agreedAt = null;
-        this.updatedAt = LocalDateTime.now();
     }
 }
